change type of fcheck arg to double

+++ trunk/ptclassic/src/kernel/Display.h
@@ -62,7 +62,7 @@ public:
         void terminate();
 
 private:
-	void fcheck(float);
+	void fcheck(double);
 	Block* blockIamIn;
 	const char* opt;
 	const char* sf;
